WeatherData1027

+++ app/src/main/java/kr/or/mrhi/openweather/Data/FeelsLike.kt
@@ -2,7 +2,7 @@ package kr.or.mrhi.openweather.Data
 
 data class FeelsLike(
     val day: Double,
-    val eve: Int,
+    val eve: Double,
     val morn: Double,
     val night: Double
 )
